+++ Changelog.md
@@ -1,8 +1,12 @@
 # Next
 - **Breaking Change** Replaced `parameters` & `parameterEncoding` in `TargetType` with extended `Task` cases.
 - **Breaking Change** Flattened `UploadType` and `DownloadType` into `Task` cases.
+- **Breaking Change** Replaced `shouldAuthorize: Bool` in `AccessTokenAuthorizable` with `authorizationType: AuthorizationType`.
+- **Breaking Change** Replaced `token` in `AccessTokenPlugin` with `tokenClosure`.
+- **Breaking Change** `TargetTypes` no longer receive the `Authorization: Bearer <token>` header by default when using `AccessTokenPlugin`.
 - Added Swift 4.0 support.
 - Added all the `filter`/`map` operators that were available for `Observable<Response>` to `Single<Response>` as well.
+- Added `AuthorizationType` to `AccessTokenAuthorizable` representing request headers of `.none`, `.basic`, and `.bearer`. 
 
 # 9.0.0-alpha.1